Actualizaciones

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class UsersController extends Controller
 {
@@ -38,16 +39,20 @@ class UsersController extends Controller
     public function store(Request $request)
     {
 
-      request()->validate([
-        'label' => 'required',
+      $data=request()->validate([
+        'name' => 'required',
+        'paternal_surname' => 'required',
+        'maternal_surname' => 'required',
+        'email' => ['required', 'email', 'unique:users'],
+        'password' => 'required',
+        'code' => ['required', 'unique:users'],
+        'role_id' => 'required',
 
     ]);
 
-      user::create([  
+      $data['password']=bcrypt(request('password'));
+      User::create($data);
 
-       'label' => request('label'),
-
-   ]);
         return redirect()->route('users.index');
     }
 
@@ -83,19 +88,30 @@ class UsersController extends Controller
     public function update(Request $request, user $user)
 
     {
+        $data=request()->validate([
+        'name' => 'required',
+        'paternal_surname' => 'required',
+        'maternal_surname' => 'required',
+        'email' => [
+            'required',
+             Rule::unique('users')->ignore($user->id),
+                    ],
+        'password' =>'',
+        'code' => ['required', Rule::unique('users')->ignore($user->id),],
+        'role_id' => 'required',
 
-      request()->validate([
-        'label' => 'required',
+    ]);
 
-        ]);
+    if (request ('password')==null) {
+         unset ($data ['password']);
+    }
 
+    else 
+     {
+        $data['password']=bcrypt(request('password'));
+     } 
 
-        $user->update([  
-
-            'label' => request('label'),
-
-        ]);
-
+        $user->update ($data);
          return redirect()->route('users.index');
     }
 
@@ -105,8 +121,12 @@ class UsersController extends Controller
      * @param  \App\Models\Career  $career
      * @return \Illuminate\Http\Response
      */
+    
     public function destroy(user $user)
     {
+        if ($user->role_id==User::ADMIN && User::where('role_id', User::ADMIN)->count() ==1) {
+         return redirect()->back();
+        }
         $user->delete();
 
         return redirect()->route('users.index');
